magento/adobe-stock-integration#1614: [Refactoring] PHP - refactored php files
