+++ MediaGalleryCatalogIntegration/README.md
@@ -1,0 +1,3 @@
+# Magento_MediaGalleryCatalogIntegration
+
+The purpose of this module is for extending catalog image uploader functionality.
