+++ MediaGalleryCatalogIntegration/etc/adminhtml/di.xml
@@ -1,0 +1,12 @@
+<?xml version="1.0"?>
+<!--
+/**
+ * Copyright © Magento, Inc. All rights reserved.
+ * See COPYING.txt for license details.
+ */
+-->
+<config xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:noNamespaceSchemaLocation="urn:magento:framework:ObjectManager/etc/config.xsd">
+    <type name="Magento\Catalog\Model\ImageUploader">
+        <plugin name="save_category_image" type="Magento\MediaGalleryCatalogIntegration\Plugin\SaveBaseCategoryImageInformation"/>
+    </type>
+</config>
