+++ MediaGalleryCatalogIntegration/modman
@@ -1,0 +1,1 @@
+./ app/code/Magento/MediaGalleryCatalogIntegration
