+++ MediaGalleryIntegration/etc/adminhtml/di.xml
@@ -14,7 +14,4 @@
     <type name="Magento\Framework\File\Uploader">
         <plugin name="save_asset_image" type="Magento\MediaGalleryIntegration\Plugin\SaveImageInformation"/>
     </type>
-    <type name="Magento\Catalog\Model\ImageUploader">
-        <plugin name="save_category_image" type="Magento\MediaGalleryIntegration\Plugin\SaveBaseCategoryImageInformation"/>
-    </type>
 </config>
