+++ MediaGallerySynchronization/Model/CreateAssetFromFile.php
@@ -7,19 +7,12 @@ declare(strict_types=1);
 
 namespace Magento\MediaGallerySynchronization\Model;
 
-use Magento\Framework\App\Filesystem\DirectoryList;
-use Magento\Framework\Exception\FileSystemException;
 use Magento\Framework\Exception\LocalizedException;
 use Magento\Framework\Exception\ValidatorException;
-use Magento\Framework\Filesystem;
-use Magento\Framework\Filesystem\Directory\Read;
-use Magento\Framework\Filesystem\Driver\File;
 use Magento\MediaGalleryApi\Api\Data\AssetInterface;
 use Magento\MediaGalleryApi\Api\Data\AssetInterfaceFactory;
 use Magento\MediaGalleryApi\Api\GetAssetsByPathsInterface;
-use Magento\MediaGalleryMetadataApi\Api\Data\MetadataInterface;
 use Magento\MediaGalleryMetadataApi\Api\ExtractMetadataInterface;
-use Magento\MediaGallerySynchronizationApi\Model\GetContentHashInterface;
 
 /**
  * Create media asset object based on the file information
@@ -28,24 +21,9 @@ use Magento\MediaGallerySynchronizationApi\Model\GetContentHashInterface;
 class CreateAssetFromFile
 {
     /**
-     * @var Filesystem
-     */
-    private $filesystem;
-
-    /**
-     * @var File
-     */
-    private $driver;
-
-    /**
      * @var GetAssetsByPathsInterface
      */
     private $getMediaGalleryAssetByPath;
-
-    /**
-     * @var Read
-     */
-    private $mediaDirectory;
 
     /**
      * @var AssetInterfaceFactory
@@ -53,9 +31,9 @@ class CreateAssetFromFile
     private $assetFactory;
 
     /**
-     * @var GetContentHashInterface
+     * @var UpdateAsset
      */
-    private $getContentHash;
+    private $updateAsset;
 
     /**
      * @var ExtractMetadataInterface
@@ -63,26 +41,20 @@ class CreateAssetFromFile
     private $extractMetadata;
 
     /**
-     * @param Filesystem $filesystem
      * @param AssetInterfaceFactory $assetFactory
-     * @param File $driver
      * @param GetAssetsByPathsInterface $getMediaGalleryAssetByPath
-     * @param GetContentHashInterface $getContentHash
+     * @param UpdateAsset $updateAsset
      * @param ExtractMetadataInterface $extractMetadata
      */
     public function __construct(
-        Filesystem $filesystem,
         AssetInterfaceFactory $assetFactory,
-        File $driver,
         GetAssetsByPathsInterface $getMediaGalleryAssetByPath,
-        GetContentHashInterface $getContentHash,
+        UpdateAsset $updateAsset,
         ExtractMetadataInterface $extractMetadata
     ) {
-        $this->filesystem = $filesystem;
         $this->assetFactory = $assetFactory;
-        $this->driver = $driver;
         $this->getMediaGalleryAssetByPath = $getMediaGalleryAssetByPath;
-        $this->getContentHash = $getContentHash;
+        $this->updateAsset = $updateAsset;
         $this->extractMetadata = $extractMetadata;
     }
 
@@ -97,47 +69,11 @@ class CreateAssetFromFile
     public function execute(\SplFileInfo $file): AssetInterface
     {
         $path = $file->getPath() . '/' . $file->getFileName();
-
-        [$width, $height] = getimagesize($path);
         $asset = $this->getAsset($path);
-
         $metadata = $this->extractMetadata->execute($path);
+        $updatedAsset = $this->updateAsset->execute($file, $asset, $metadata);
 
-        return $this->assetFactory->create(
-            [
-                'id' => $asset ? $asset->getId() : null,
-                'path' => $this->getRelativePath($path),
-                'title' => $this->getAssetTitle($file, $asset, $metadata),
-                'description' => $metadata->getDescription(),
-                'createdAt' => $asset ? $asset->getCreatedAt() :
-                    (new \DateTime())->setTimestamp($file->getCTime())->format('Y-m-d H:i:s'),
-                'updatedAt' => (new \DateTime())->setTimestamp($file->getMTime())->format('Y-m-d H:i:s'),
-                'width' => $width,
-                'height' => $height,
-                'hash' => $this->getHashImageContent($path),
-                'size' => $file->getSize(),
-                'contentType' => $asset ? $asset->getContentType() : 'image/' . $file->getExtension(),
-                'source' => $asset ? $asset->getSource() : 'Local'
-            ]
-        );
-    }
-
-    /**
-     * Returns asset title from metadata if available
-     *
-     * @param \SplFileInfo $file
-     * @param null|AssetInterface $asset
-     * @param MetadataInterface $metadata
-     */
-    private function getAssetTitle(\SplFileInfo $file, ?AssetInterface $asset, MetadataInterface $metadata): string
-    {
-        $title = $file->getBasename('.' . $file->getExtension());
-        if ($asset) {
-            $title = $asset->getTitle();
-        } elseif ($metadata->getTitle() !== "") {
-            $title = $metadata->getTitle();
-        }
-        return $title;
+        return $updatedAsset;
     }
 
     /**
@@ -150,54 +86,7 @@ class CreateAssetFromFile
      */
     private function getAsset(string $path): ?AssetInterface
     {
-        $asset = $this->getMediaGalleryAssetByPath->execute([$this->getRelativePath($path)]);
+        $asset = $this->getMediaGalleryAssetByPath->execute([$this->updateAsset->getRelativePath($path)]);
         return !empty($asset) ? $asset[0] : null;
-    }
-
-    /**
-     * Get correct path for media asset
-     *
-     * @param string $file
-     * @return string
-     * @throws ValidatorException
-     */
-    private function getRelativePath(string $file): string
-    {
-        $path = $this->getMediaDirectory()->getRelativePath($file);
-
-        if ($this->driver->getParentDirectory($path) === '.') {
-            $path = '/' . $path;
-        }
-
-        return $path;
-    }
-
-    /**
-     * Get hash image content.
-     *
-     * @param string $path
-     * @return string
-     * @throws ValidatorException
-     * @throws FileSystemException
-     */
-    private function getHashImageContent(string $path): string
-    {
-        $mediaDirectory = $this->getMediaDirectory();
-        $imageDirectory = $mediaDirectory->readFile($mediaDirectory->getRelativePath($path));
-        $hashedImageContent = $this->getContentHash->execute($imageDirectory);
-        return $hashedImageContent;
-    }
-
-    /**
-     * Retrieve media directory instance with read permissions
-     *
-     * @return Read
-     */
-    private function getMediaDirectory(): Read
-    {
-        if (!$this->mediaDirectory) {
-            $this->mediaDirectory = $this->filesystem->getDirectoryRead(DirectoryList::MEDIA);
-        }
-        return $this->mediaDirectory;
     }
 }
